Refactor out Configurable
