+++ lib/ladder.rb
@@ -2,7 +2,8 @@ require 'ladder/version'
 require 'ladder/config'
 
 module Ladder
-  autoload :File,       'ladder/file'
-  autoload :Resource,   'ladder/resource'
-  autoload :Searchable, 'ladder/searchable'
+  autoload :File,         'ladder/file'
+  autoload :Resource,     'ladder/resource'
+  autoload :Searchable,   'ladder/searchable'
+  autoload :Configurable, 'ladder/configurable'
 end
