+++ lib/ladder/resource.rb
@@ -10,6 +10,7 @@ module Ladder
 
     include Mongoid::Document
     include ActiveTriples::Identifiable
+    include Ladder::Configurable
     include Ladder::Resource::Serializable
 
     included { configure_model }
@@ -280,24 +281,7 @@ module Ladder
         new_object
       end
 
-      ##
-      # Set a default base URI based on the Ladder::Config settings
-      #
-      # @return [RDF::URI]
-      def base_uri
-        RDF::URI.new(Ladder::Config.settings[:base_uri]) / name.underscore.pluralize
-      end
-
       protected
-
-      ##
-      # Register with Ladder and set the default base URI
-      #
-      # @return [void]
-      def configure_model
-        configure base_uri: base_uri
-        Ladder::Config.register_model self unless Ladder::Config.models.include? self
-      end
 
       ##
       # Propagate base URI and properties to subclasses
